long overdue refactor out of the various agent / model classes into their own dirs
