+++ vegetation/model/vegetation.py
@@ -1,0 +1,226 @@
+import mesa
+import mesa_geo as mg
+import numpy as np
+import shapely.geometry as sg
+from shapely.ops import transform
+import json
+import logging
+
+from vegetation.config.stages import LifeStage
+from vegetation.space.veg_cell import StudyArea, VegCell
+from vegetation.utils.spatial import transform_point_wgs84_utm
+from vegetation.config.paths import INITIAL_AGENTS_PATH
+from vegetation.config.logging import (
+    LogConfig,
+    SimLogger,
+    SimEventType,
+)
+from vegetation.utils.zarr_manager import (
+    get_array_from_nested_cell_list,
+)
+
+
+class Vegetation(mesa.Model):
+
+    @property
+    def sim_logger(self):
+        if not hasattr(self, "_sim_logger"):
+            self._sim_logger = SimLogger()
+        return self._sim_logger
+
+    def __init__(
+        self,
+        bounds,
+        export_data=False,
+        num_steps=20,
+        management_planting_density=0.01,
+        epsg=4326,
+        log_config_path=None,
+        log_level=None,
+    ):
+        super().__init__()
+
+        # Initialize logging config first
+        if log_config_path:
+            LogConfig.initialize(log_config_path)
+
+        # To get this set up, assume sim has logging.INFO level
+        self.log_level = logging.INFO
+
+        self.bounds = bounds
+        self.num_steps = num_steps
+        self.management_planting_density = management_planting_density
+        self._on_start_executed = False
+        self.sim_idx = 1
+
+        # mesa setup
+        self.space = StudyArea(bounds, epsg=epsg, model=self)
+        self.datacollector = mesa.DataCollector(
+            {
+                "Mean Age": "mean_age",
+                "N Agents": "n_agents",
+                "N Seeds": "n_seeds",
+                "N Seedlings": "n_seedlings",
+                "N Juveniles": "n_juveniles",
+                "N Adults": "n_adults",
+                "N Breeding": "n_breeding",
+                "% Refugia Cells Occupied": "pct_refugia_cells_occupied",
+            }
+        )
+
+        self.life_stage_zarr = None
+
+    def _initialize_zarr(self):
+
+        self.life_stage_zarr = initialize_synchronized_zarr(
+            (self.space.raster_layer._width, self.space.raster_layer._height),
+            ".life_stage.zarr",
+            crs=self.space.crs.to_string(),
+            transformer_json=self.space._transformer.to_json(),
+        )
+
+    def _append_current_timestep_to_zarr(self):
+        timestep_array = get_array_from_nested_cell_list(self.space.raster_layer.cells)
+        append_synchronized_timestep(
+            zarr_array=self.life_stage_zarr,
+            sim_idx=self.sim_idx,
+            timestep_idx=self.steps,
+            timestep_array=timestep_array,
+        )
+
+    def _on_start(self):
+
+        self.sim_logger.log_sim_event(self, SimEventType.ON_START)
+
+        self.space.get_elevation()
+        self.space.get_aridity()
+        self.space.get_refugia_status()
+
+        with open(INITIAL_AGENTS_PATH, "r") as f:
+            initial_agents_geojson = json.loads(f.read())
+
+        self._add_agents_from_geojson(initial_agents_geojson)
+
+        self._initialize_zarr()
+
+        self._on_start_executed = True
+
+    def _add_agents_from_geojson(self, agents_geojson):
+        agents = mg.AgentCreator(JoshuaTreeAgent, model=self).from_GeoJSON(
+            agents_geojson
+        )
+
+        # TODO: Find a way to update life stage on init
+        # Issue URL: https://github.com/SchmidtDSE/mesa_abm_poc/issues/9
+        # Since .from_GeoJSON() sets attributes after init, we call
+        # _update_life_stage after init, but before we add to the grid
+        self.agents.select(agent_type=JoshuaTreeAgent).do("_update_life_stage")
+
+        self.space.add_agents(agents)
+        self.update_metrics()
+
+    # def add_agents_from_management_draw(event, geo_json, action):
+    def add_agents_from_management_draw(self, *args, **kwargs):
+
+        assert kwargs.get("action") == "create"
+        management_area = kwargs.get("geo_json")
+
+        outplanting_point_locations = self._generate_planting_points(management_area)
+
+        self.sim_logger.log_sim_event(
+            self,
+            SimEventType.ON_MANAGE,
+            context={"n_agents": len(outplanting_point_locations)},
+        )
+
+        for management_x_wgs84, management_y_wgs84 in outplanting_point_locations:
+
+            # TODO: Vegetation model doesn't know its own CRS
+            # Issue URL: https://github.com/SchmidtDSE/mesa_abm_poc/issues/26
+            management_agent = JoshuaTreeAgent(
+                model=self,
+                geometry=sg.Point(management_x_wgs84, management_y_wgs84),
+                crs="EPSG:4326",
+                age=20,
+                parent_id=None,
+            )
+            management_agent._update_life_stage()
+
+            self.space.add_agents(management_agent)
+
+    def _generate_planting_points(self, geo_json):
+        # Convert GeoJSON to Shapely polygon
+        coords = geo_json[0]["geometry"]["coordinates"][0]
+        polygon = sg.Polygon(coords)
+
+        # Get UTM zone from polygon centroid
+        lon, lat = polygon.centroid.x, polygon.centroid.y
+        wgs84_to_utm, utm_to_wgs84 = transform_point_wgs84_utm(lon, lat)
+
+        # Project polygon to UTM
+        utm_polygon = transform(wgs84_to_utm.transform, polygon)
+        area = utm_polygon.area
+        num_points = int(area * self.management_planting_density)
+
+        points = []
+        minx, miny, maxx, maxy = utm_polygon.bounds
+
+        while len(points) < num_points:
+            x_utm = np.random.uniform(minx, maxx)
+            y_utm = np.random.uniform(miny, maxy)
+            point_utm = sg.Point(x_utm, y_utm)
+
+            if utm_polygon.contains(point_utm):
+                management_x_wgs84, management_y_wgs84 = utm_to_wgs84.transform(
+                    x_utm, y_utm
+                )
+                points.append((management_x_wgs84, management_y_wgs84))
+
+        return points
+
+    def update_metrics(self):
+        # Mean age
+        mean_age = self.agents.select(agent_type=JoshuaTreeAgent).agg("age", np.mean)
+        self.mean_age = mean_age
+
+        # Number of agents by life stage
+        count_dict = (
+            self.agents.select(agent_type=JoshuaTreeAgent).groupby("life_stage").count()
+        )
+        self.n_seeds = count_dict.get(LifeStage.SEED, 0)
+        self.n_seedlings = count_dict.get(LifeStage.SEEDLING, 0)
+        self.n_juveniles = count_dict.get(LifeStage.JUVENILE, 0)
+        self.n_adults = count_dict.get(LifeStage.ADULT, 0)
+        self.n_breeding = count_dict.get(LifeStage.BREEDING, 0)
+        self.n_dead = count_dict.get(LifeStage.DEAD, 0)
+
+        # Number of agents (JoshuaTreeAgent)
+        n_agents = len(self.agents.select(agent_type=JoshuaTreeAgent))
+        self.n_agents = n_agents - self.n_dead
+
+        # Number of refugia cells occupied by JoshuaTreeAgents
+        count_dict = (
+            self.agents.select(agent_type=VegCell)
+            .select(filter_func=lambda agent: agent.refugia_status)
+            .groupby("occupied_by_jotr_agents")
+            .count()
+        )
+        self.pct_refugia_cells_occupied = count_dict.get(True, 0) / (
+            count_dict.get(True, 0) + count_dict.get(False, 0)
+        )
+
+    def step(self):
+
+        if not self._on_start_executed:
+            self._on_start()
+
+        self.sim_logger.log_sim_event(self, SimEventType.ON_STEP)
+
+        self._append_current_timestep_to_zarr()
+
+        # Step agents
+        self.agents.shuffle_do("step")
+        self.update_metrics()
+
+        # Collect data
+        self.datacollector.collect(self)
