+++ vegetation/utils/spatial.py
